Add java doc for the interface classes in the messaging package. Fixed compilation issues. Removed modules that are not completed and compiling. Fixed the pom files.

+++ source-java/messaging/src/main/java/org/eris/transport/io/IoNetworkConnection.java
@@ -50,7 +50,7 @@ public class IoNetworkConnection implements NetworkConnection<ByteBuffer>
         try
         {
             InetAddress address = InetAddress.getByName(_settings.getHost());
-            _socket.connect(new InetSocketAddress(address, _settings.getPort()), _settings.getConnectTimeout());
+            _socket.connect(new InetSocketAddress(address, _settings.getPort()), (int)_settings.getConnectTimeout());
         }
         catch (UnknownHostException e)
         {
